# api/openapi.yaml
openapi: 3.1.0
info:
  title: Simple-VPN-Builder Control Plane API
  version: 0.1.0v
  description: High-performance VPN orchestration and client protocol provisioning control plane.
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /sub/{token}:
    get:
      summary: Universal Client Subscription Delivery
      description: Returns multi-format subscription profile (Base64 URIs, Sing-box JSON v1.8+, Clash Meta YAML) with Subscription-Userinfo quota headers.
      operationId: getSubscription
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [base64, singbox, sing-box, clash, clash-meta, mihomo, wireguard, wg, amneziawg, awg, amnezia, json]
      responses:
        '200':
          description: Subscription payload formatted for the detected or requested client
          headers:
            Subscription-Userinfo:
              schema:
                type: string
                example: "upload=1048576; download=5242880; total=107374182400; expire=1735689600"
            Profile-Update-Interval:
              schema:
                type: string
                example: "24"
            Profile-Title:
              schema:
                type: string
                example: "Simple-VPN-Network"
            Content-Disposition:
              schema:
                type: string
                example: "attachment; filename=\"subscription.yaml\""
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
            application/json:
              schema:
                type: object
            application/x-yaml:
              schema:
                type: string
        '400':
          description: Invalid subscription token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: User account inactive, expired, or data limit exceeded
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/login:
    post:
      summary: Administrator Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Invalid credentials or missing 2FA code
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/refresh:
    post:
      summary: Refresh Access Token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or revoked refresh token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/logout:
    post:
      summary: Invalidate Session and Revoke Token
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: logged_out
  /api/v1/auth/totp/setup:
    post:
      summary: Provision TOTP 2FA Secret
      operationId: setupTOTP
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TOTP secret and provisioning URI generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TOTPSetupResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/totp/verify:
    post:
      summary: Verify Code and Activate TOTP
      operationId: verifyTOTP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPVerifyRequest'
      responses:
        '200':
          description: Two-factor authentication successfully enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: totp_enabled
        '400':
          description: Invalid passcode or secret
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/auth/change-password:
    post:
      summary: Rotate Credentials With Old Password Proof
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password rotated
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Invalid old password
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/billing/stars/confirm:
    post:
      summary: Confirm Telegram Stars Payment
      operationId: confirmStarsPayment
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Stars order closed
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /admins:
    get:
      tags: [Admins]
      summary: List administrators
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of admins
    post:
      tags: [Admins]
      summary: Create administrator
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Admin created
  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke API key
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: API key revoked
  /ai/chat:
    post:
      tags: [AI]
      summary: Chat with AI copilot
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Chat completion
  /ai/actions/{token}/execute:
    post:
      tags: [AI]
      summary: Execute approved AI action
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Action executed
  /ai/settings:
    get:
      tags: [AI]
      summary: Get AI settings
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: AI settings
    post:
      tags: [AI]
      summary: Update AI settings
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: AI settings updated
  /ai/test:
    post:
      tags: [AI]
      summary: Test AI provider connection
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Connection test result
  /nodes:
    get:
      tags: [Nodes]
      summary: List VPN nodes
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Paginated list of nodes
    post:
      tags: [Nodes]
      summary: Register a new node
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Node created successfully
  /nodes/{id}:
    get:
      tags: [Nodes]
      summary: Get node details
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Node details
    patch:
      tags: [Nodes]
      summary: Update node
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Node updated
    delete:
      tags: [Nodes]
      summary: Delete node
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Node deleted
  /users:
    get:
      tags: [Users]
      summary: List VPN users
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Paginated list of users
    post:
      tags: [Users]
      summary: Create VPN user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: User created
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user details
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
    patch:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User updated
    delete:
      tags: [Users]
      summary: Delete user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
  /plans:
    get:
      tags: [Plans]
      summary: List subscription plans
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of plans
    post:
      tags: [Plans]
      summary: Create subscription plan
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Plan created
  /credentials:
    get:
      tags: [Credentials]
      summary: List credentials
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of credentials
    post:
      tags: [Credentials]
      summary: Provision credential
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Credential provisioned
...
  /analytics/overview:
    get:
      tags: [Analytics]
      summary: Global traffic overview
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Aggregate traffic overview
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
          example: supersecret
        totp_code:
          type: string
          example: "123456"
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        invalid_params:
          type: object
          additionalProperties:
            type: string